+++ app/src/main/java/nextstep/jwp/dashboard/exception/DuplicateUserException.java
@@ -1,5 +1,7 @@
 package nextstep.jwp.dashboard.exception;
 
+import nextstep.jwp.web.exception.BadRequestException;
+
 public class DuplicateUserException extends BadRequestException {
     public DuplicateUserException(String account) {
         super(String.format("Duplicated user information. account : %s", account));
