+++ app/src/main/java/nextstep/jwp/dashboard/exception/UserNotFoundException.java
@@ -1,5 +1,7 @@
 package nextstep.jwp.dashboard.exception;
 
+import nextstep.jwp.web.exception.NotFoundException;
+
 public class UserNotFoundException extends NotFoundException {
     public UserNotFoundException(String account) {
         super(String.format("User does not exist. account : %s", account));
